tutorial/supply/pyproject.toml
[tool.poetry]
name = "supply"
version = "0.1.0"
description = ""
authors = ["chengzhenbo <50646861+chengzhenbo@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.3"
pandera = "^0.16.1"
openpyxl = "^3.1.2"
pypinyin = "^0.49.0"
fastapi = "^0.100.1"
uvicorn = "^0.23.1"
python-multipart = "^0.0.6"
alembic = "^1.11.1"
fastapi-sqlalchemy = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
